Check type before construct ConstantExpression
Move the type check to construction time so that we will not create
expression that will eventually fail in serialization.

+++ presto-main/src/main/java/com/facebook/presto/sql/planner/RowExpressionInterpreter.java
@@ -547,7 +547,7 @@ public class RowExpressionInterpreter
                                 DEREFERENCE,
                                 node.getType(),
                                 toRowExpression(base, node.getArguments().get(0)),
-                                toRowExpression(index, node.getArguments().get(1)));
+                                toRowExpression((long) index, node.getArguments().get(1)));
                     }
                     return interpretDereference(base, node.getType(), index);
                 }
